+++ lib/src/entities/address.dart
@@ -1,0 +1,6 @@
+class Address {
+  String name;
+  String mailAddress;
+
+  Address([this.mailAddress, this.name]);
+}