+++ lib/src/entities/send_report.dart
@@ -1,0 +1,10 @@
+import 'message.dart';
+import 'problem.dart';
+
+class SendReport {
+  final Message mail;
+  final bool sent;
+  final List<Problem> validationProblems;
+
+  SendReport(this.mail, this.sent, {this.validationProblems});
+}